tests/wpt/tests/css/css-scroll-snap/parsing/scroll-snap-stop-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-snap-stop with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-snap-stop">
<meta name="assert" content="scroll-snap-stop supports the full grammar 'normal | always'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-snap-stop", "normal");
test_valid_value("scroll-snap-stop", "always");
</script>
</body>
</html>
